content-bot-complete/requirements.txt
openai==0.28.1
tweepy==4.14.0
praw==7.7.0
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
mailchimp3==3.0.21
sendgrid==6.10.0
sqlalchemy==2.0.23
flask==3.0.0
gunicorn==21.2.0
